Add handling of long file lines
When a line in a text file is too long, currently only part of it is
read and the rest is skipped quietly. With this change if the line is
too long, an exception is thrown. This way both the user and the
developer could know that there is something wrong and maybe the results
can not be counted on.